commit 10

+++ csv_parser/__init__.py
@@ -8,6 +8,8 @@ import urllib.request
 from os import getcwd, chdir, mkdir
 from os import chdir
 import os
+import os.path
+
 rep_cour = os.getcwd()
 print(rep_cour)
 chdir("csv_uploads")
@@ -16,29 +18,19 @@ print(rep_cour)
 
 # recuprer la page avec request.get qui lance une requete (http/https)
 for url in ['https://people.sc.fsu.edu/~jburkardt/data/csv/addresses.csv', 'https://people.sc.fsu.edu/~jburkardt/data/csv/airtravel.csv','https://people.sc.fsu.edu/~jburkardt/data/csv/biostats.csv']:
-    try:
+   
         # telecharegr le fichier et le stocker dans 
         chaine = url.split("/")
-        name_file = chaine[len(chaine) - 1] 
-        #name_files = r"C:\Users\soray\Documents\cours python\Projets\csv-parser\csv_parser\csv_uploads"
-       
-        #if not os.path.isfile(name_files){
-    
-        with open(name_file, 'wb') as file:
-            # get request
-            response = requests.get(url) # response est un objet qui permet d'inspecter le resultat de la requete 
-            # write to file
-            file.write(response.content)
-            # si response est successful, no Exception will be raised 
-            response.raise_for_status() # un HTTPErrorsera déclenché pour certains codes d'état si le code d'etat indique reussit il continu 
-           # }
-           
-       
-    except HTTPError as http_err:
-                print(f'HTTP error occurred: {http_err}')   
-    except Exception as err:
-        print(f'Other error occurred: {err}')  
-    else:
-        print('Success!')
-
-os.path.exists('/Users/toto/Desktop')
+        name_file = chaine[len(chaine) - 1]        
+        if os.path.isfile(name_file):
+            print('le fichier existe ')
+            
+        else:
+            print('le fichier n existe pas')
+            with open(name_file, 'wb') as file:
+                # get request
+                response = requests.get(url)  # response est un objet qui permet d'inspecter le resultat de la requete 
+                # write to file
+                file.write(response.content)
+                # si response est successful, no Exception will be raised 
+                response.raise_for_status() # un HTTPErrorsera déclenché pour certains codes d'état si le code d'etat indique reussit il continu
